+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/ai/goals/target/FindAllTargetsGoal.java
@@ -1,4 +1,4 @@
-package net.picopress.mc.mods.zombietactics2.goals.target;
+package net.picopress.mc.mods.zombietactics2.ai.goals.target;
 
 import net.picopress.mc.mods.zombietactics2.util.Tactics;
 import net.picopress.mc.mods.zombietactics2.impl.Plane;
@@ -20,14 +20,12 @@ import org.jetbrains.annotations.Nullable;
 
 import oshi.util.tuples.Pair;
 
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.List;
+import java.util.*;
 
 
 // the new improved target finding goal
 public class FindAllTargetsGoal extends TargetGoal {
-    public static final List<Pair<LivingEntity, Path>> cache_path = new ArrayList<>();
+    public static final Map<LivingEntity, Path> cache_path = new HashMap<>();
     private final List<Pair<Class<? extends LivingEntity>, Integer>> list;
     private final Plane plane;
     private List<LivingEntity> imposters;
@@ -111,19 +109,11 @@ public class FindAllTargetsGoal extends TargetGoal {
                 int idx = 0;
 
                 if(Config.findTargetType == FindTargetType.INTENSIVE) {
-                    Path path = null;
-                    boolean found = false;
-                    for(var p: cache_path) {
-                        if(p.getA() == amogus) {
-                            path = p.getB();
-                            found = true;
-                            break;
-                        }
-                    }
-                    if(!found) {
+                    Path path = cache_path.get(amogus);
+                    if(path == null) {
                         // use cache to prevent overloading
                         path = mob.getNavigation().createPath(amogus, Config.accuracy);
-                        cache_path.add(new Pair<>(amogus, path));
+                        cache_path.put(amogus, path);
                     }
                     if(path != null) {
                         score += path.getNodeCount();
@@ -156,7 +146,7 @@ public class FindAllTargetsGoal extends TargetGoal {
                 // idx must match the target list unless priorities are invalid
                 score *= list.get(idx).getB();
 
-                // getting insane
+                // getting crazy
                 if(mob.hasLineOfSight(amogus)) score /= 2;
                 if(delta.getY() >= -2) score /= 2;
                 score *= Tactics.Heuristic.getEnemyPower(amogus);