+++ common/src/main/java/net/picopress/mc/mods/zombietactics2/mixin/ServerLevelMixin.java
@@ -2,7 +2,7 @@ package net.picopress.mc.mods.zombietactics2.mixin;
 
 import net.picopress.mc.mods.zombietactics2.Config;
 import net.picopress.mc.mods.zombietactics2.attachments.FindTargetType;
-import net.picopress.mc.mods.zombietactics2.goals.target.FindAllTargetsGoal;
+import net.picopress.mc.mods.zombietactics2.ai.goals.target.FindAllTargetsGoal;
 
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ServerLevel;
@@ -11,6 +11,7 @@ import net.minecraft.core.RegistryAccess;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.dimension.DimensionType;
 import net.minecraft.world.level.storage.WritableLevelData;
+import net.minecraft.world.entity.LivingEntity;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
@@ -18,6 +19,8 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.function.BooleanSupplier;
 
 
@@ -32,24 +35,18 @@ public abstract class ServerLevelMixin extends Level {
     // garbage
     @Inject(method = "tick", at = @At("TAIL"))
     public void tick(BooleanSupplier hasTimeLeft, CallbackInfo ci) {
-        // run if FindTargetType is not LINEAR
+        // run if FindTargetType is INTENSIVE
         if(Config.findTargetType == FindTargetType.INTENSIVE) {
             ++ zombie_tactics$duration;
-            if(zombie_tactics$duration > 20) {
+            if(zombie_tactics$duration > 22) {
                 zombie_tactics$duration = 0;
-                while(true) {
-                    boolean mark = true;
-                    int idx = 0;
-                    for(var cp: FindAllTargetsGoal.cache_path) {
-                        if(!cp.getA().isAlive()) {
-                            mark = false;
-                            break;
-                        }
-                        ++ idx;
+                List<LivingEntity> dead = new ArrayList<>();
+                for(LivingEntity x: FindAllTargetsGoal.cache_path.keySet()) {
+                    if(!x.isAlive()) {
+                        dead.add(x);
                     }
-                    if(mark) break;
-                    FindAllTargetsGoal.cache_path.remove(idx);
                 }
+                dead.forEach(FindAllTargetsGoal.cache_path::remove);
             }
         }
     }