+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-- support for the new version 1.21.7
+# Zombie Tactics 2.2 Changelog
+- Zombies can get off a boat
+- Zombies can swim in the water
+- Zombies will kill ambient animals
+- Many bug fixes